Refactorisation et ajout de l'intégralité des tests unitaires. (1/2)

+++ api/src/main/java/com/discoodle/api/service/RoomService.java
@@ -5,8 +5,10 @@ import com.discoodle.api.repository.MessagesRepository;
 import com.discoodle.api.repository.RoomRepository;
 import com.discoodle.api.repository.UserRepository;
 import lombok.AllArgsConstructor;
+import org.aspectj.weaver.ast.Test;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -26,7 +28,7 @@ public class RoomService {
         );
         Room finalRoom = roomRepository.save(room);
         for (Long room_member : room_members) {
-            if (userRepository.existsById(room_member)) {
+            if (userRepository.existsById(room_member) && userRepository.existsById(room_member)) {
                 roomRepository.addNewMember(finalRoom.getRoom_id(), room_member);
             }
         }
@@ -42,9 +44,10 @@ public class RoomService {
     }
 
     public Optional<User> removeMember(String room_id, Long user_id) {
-        if (roomRepository.existsById(room_id) && userRepository.existsById(user_id)) {
+        Optional<User> test = userRepository.findById(user_id);
+        if (roomRepository.existsById(room_id) && test.isPresent()) {
             Optional<Room> room = roomRepository.findById(room_id);
-            if (room.get().getUsers().size() > 1) {
+            if (room.get().getUsers().size() > 1 && room.get().getUsers().contains(test.get())) {
                 roomRepository.removeMember(room_id, user_id);
                 if (user_id.equals(room.get().getRoom_admin())) {
                     if (room.get().getUsers().get(1).getId().equals(user_id))
@@ -69,29 +72,36 @@ public class RoomService {
         return messagesRepository.getMessagesOfUser(user_id);
     }
 
-    public void addNewMember(String room_id, Long user_id) {
+    public Optional<Room> addNewMember(String room_id, Long user_id) {
         Optional<User> test = userRepository.findById(user_id);
         if (test.isPresent() && !roomRepository.findById(room_id).get().getUsers().contains(test.get())) {
             roomRepository.addNewMember(room_id, user_id);
         }
+        return Optional.empty();
     }
 
     public Optional<Room> changeLinkPicture(String room_id, String link_to_avatar) {
-        if (roomRepository.changeLinkPicture(room_id, link_to_avatar) == 1) {
-            return roomRepository.findById(room_id);
+        Optional<Room> test = roomRepository.findById(room_id);
+        if (test.isPresent() && roomRepository.changeLinkPicture(room_id, link_to_avatar) == 1) {
+            return test;
         }
         return Optional.empty();
     }
 
     public Optional<Room> changeAdmin(String room_id, Long room_admin) {
-        if (roomRepository.changeAdmin(room_id, room_admin) == 1) {
-            return roomRepository.findById(room_id);
+        Optional<Room> test = roomRepository.findById(room_id);
+        if (test.isPresent() && roomRepository.changeAdmin(room_id, room_admin) == 1) {
+            return test;
         }
         return Optional.empty();
     }
 
-    public void renameRoom(String room_id, String new_name) {
-        roomRepository.renameRoom(room_id, new_name);
+    public Optional<Room> renameRoom(String room_id, String new_name) {
+        Optional<Room> test = roomRepository.findById(room_id);
+        if(test.isPresent()) {
+            roomRepository.renameRoom(room_id, new_name);
+        }
+        return Optional.empty();
     }
 
 }
